# apps/backend/swagger-output/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/catalog/upload-image": {
      "post": {
        "operationId": "CatalogController_uploadImage",
        "summary": "Upload une image de produit depuis le connector",
        "description": "Endpoint appelé par le script exécuté dans le connector pour uploader les images des produits. Le clientId est extrait du token JWT pour des raisons de sécurité. Accepte les données en base64 (JSON) depuis nodeFetch.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Image uploadée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "url": "https://files-flemme.bedones.com/whatsapp-agent/237697020290/catalog/images/849641504281228/25095720553426064-0.jpg"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erreur lors de l'upload"
          },
          "401": {
            "description": "Token invalide ou expiré"
          }
        },
        "tags": ["Catalog"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/catalog/upload-avatar": {
      "post": {
        "operationId": "CatalogController_uploadAvatar",
        "summary": "Upload l'avatar du compte WhatsApp",
        "description": "Endpoint appelé par le script exécuté dans le connector pour uploader l'avatar du compte. Le clientId est extrait du token JWT. Accepte les données en base64 (JSON) depuis nodeFetch.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Avatar uploadé avec succès"
          }
        },
        "tags": ["Catalog"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/catalog/save-client-info": {
      "post": {
        "operationId": "CatalogController_saveClientInfo",
        "summary": "Sauvegarde les informations du client WhatsApp",
        "description": "Endpoint appelé par le script pour sauvegarder les informations du compte WhatsApp Business.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Informations sauvegardées avec succès"
          }
        },
        "tags": ["Catalog"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/catalog/save-catalog": {
      "post": {
        "operationId": "CatalogController_saveCatalog",
        "summary": "Sauvegarde le catalogue (collections et produits)",
        "description": "Endpoint appelé par le script pour sauvegarder le catalogue complet avec les collections et produits.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Catalogue sauvegardé avec succès"
          }
        },
        "tags": ["Catalog"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/catalog/delete-images": {
      "post": {
        "operationId": "CatalogController_deleteImages",
        "summary": "Supprime des images obsolètes",
        "description": "Endpoint appelé par le script pour supprimer les images qui ne sont plus dans le catalogue WhatsApp. Supprime les fichiers de Minio et les entrées de la BD.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Images supprimées avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "deletedCount": 5
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erreur lors de la suppression"
          },
          "401": {
            "description": "Token invalide ou expiré"
          }
        },
        "tags": ["Catalog"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/orders": {
      "get": {
        "operationId": "OrdersController_getAllOrders",
        "summary": "Get all orders for current user",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by order status",
            "schema": {
              "enum": ["NEGOTIATION", "CONFIRMED", "DELIVERED", "CANCELLED"],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of orders retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": ["orders"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "OrdersController_createOrder",
        "summary": "Create a new order",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": ["orders"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/orders/{id}": {
      "get": {
        "operationId": "OrdersController_getOrder",
        "summary": "Get order by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Order ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Order not found"
          }
        },
        "tags": ["orders"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "OrdersController_updateOrder",
        "summary": "Update an existing order",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Order ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order updated successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Order not found"
          }
        },
        "tags": ["orders"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/orders/{id}/status": {
      "patch": {
        "operationId": "OrdersController_updateOrderStatus",
        "summary": "Update order status",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Order ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order status updated successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Order not found"
          }
        },
        "tags": ["orders"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Monorepo API",
    "description": "Interractive documentation for the Monorepo API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "conversationId": {
            "type": "string",
            "description": "Conversation ID",
            "example": "clxxx123456789"
          },
          "customerName": {
            "type": "string",
            "description": "Customer name",
            "example": "John Doe"
          },
          "customerPhone": {
            "type": "string",
            "description": "Customer phone number",
            "example": "237612345678"
          },
          "products": {
            "type": "object",
            "description": "Products (JSON array)",
            "example": [
              {
                "productId": "clxxx123",
                "name": "iPhone 15 Pro",
                "quantity": 1,
                "price": 999.99
              }
            ]
          },
          "totalAmount": {
            "type": "number",
            "description": "Total amount",
            "example": 999.99
          },
          "deliveryLocationId": {
            "type": "string",
            "description": "Delivery location ID",
            "example": "clxxx789"
          },
          "paymentMethodId": {
            "type": "string",
            "description": "Payment method ID",
            "example": "clxxx456"
          },
          "promiseDate": {
            "type": "string",
            "description": "Promise date for the order",
            "example": "2025-11-20T10:00:00Z"
          }
        },
        "required": [
          "conversationId",
          "customerName",
          "customerPhone",
          "products",
          "totalAmount"
        ]
      },
      "UpdateOrderDto": {
        "type": "object",
        "properties": {
          "customerName": {
            "type": "string",
            "description": "Customer name",
            "example": "John Doe"
          },
          "customerPhone": {
            "type": "string",
            "description": "Customer phone number",
            "example": "237612345678"
          },
          "products": {
            "type": "object",
            "description": "Products (JSON array)",
            "example": [
              {
                "productId": "clxxx123",
                "name": "iPhone 15 Pro",
                "quantity": 1,
                "price": 999.99
              }
            ]
          },
          "totalAmount": {
            "type": "number",
            "description": "Total amount",
            "example": 999.99
          },
          "deliveryLocationId": {
            "type": "string",
            "description": "Delivery location ID",
            "example": "clxxx789"
          },
          "paymentMethodId": {
            "type": "string",
            "description": "Payment method ID",
            "example": "clxxx456"
          },
          "promiseDate": {
            "type": "string",
            "description": "Promise date for the order",
            "example": "2025-11-20T10:00:00Z"
          },
          "status": {
            "type": "string",
            "description": "Order status",
            "enum": ["NEGOTIATION", "CONFIRMED", "DELIVERED", "CANCELLED"]
          }
        }
      },
      "UpdateOrderStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Order status",
            "enum": ["NEGOTIATION", "CONFIRMED", "DELIVERED", "CANCELLED"],
            "example": "CONFIRMED"
          }
        },
        "required": ["status"]
      },
...
    }
  }
}